requirements.txt
fastapi
uvicorn
sentence-transformers
qdrant-client
pydantic
datasets
pyyaml
